separated codebase into modules
