+++ src/modules/constants.js
@@ -1,0 +1,16 @@
+export const deviceTypes = {
+	"laptop-win-10": {
+		fullName: "Laptop (Windows 10)",
+		monthlyPayment: 10,
+		yearlyPayment: 100,
+		circleValue: 50
+	},
+	"laptop-win-11": {
+		fullName: "Laptop (Windows 11)",
+		monthlyPayment: 15,
+		yearlyPayment: 150,
+		circleValue: 75
+	}
+}
+
+export const postalCodesMechelen = [2800, 2801, 2811, 2812];
